+++ src/test/java/org/cicirello/search/ss/AcceptanceBandSamplingIntCostTests.java
@@ -27,8 +27,8 @@ import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.SolutionCostPair;
 import org.junit.jupiter.api.*;
 
-/** JUnit tests for AcceptanceBandSampling. */
-public class AcceptanceBandTests extends SharedTestStochasticSampler {
+/** JUnit tests for AcceptanceBandSampling for problems with int-valued costs. */
+public class AcceptanceBandSamplingIntCostTests extends SharedTestStochasticSampler {
 
   @Test
   public void testConstructorExceptions() {
@@ -107,31 +107,6 @@ public class AcceptanceBandTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCosts() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n, 20);
-      AcceptanceBandSampling<Permutation> ch = new AcceptanceBandSampling<Permutation>(h);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize();
-      assertEquals(2, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsMultipleSamples() {
     for (int n = 0; n < 10; n++) {
       IntProblem problem = new IntProblem();
@@ -150,31 +125,6 @@ public class AcceptanceBandTests extends SharedTestStochasticSampler {
       assertEquals(7, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsMultipleSamples() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n, 20);
-      AcceptanceBandSampling<Permutation> ch = new AcceptanceBandSampling<Permutation>(h);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize(5);
-      assertEquals(5, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize(2);
-      assertEquals(7, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       tracker = new ProgressTracker<Permutation>();
       ch.setProgressTracker(tracker);
       assertTrue(tracker == ch.getProgressTracker());
@@ -206,30 +156,6 @@ public class AcceptanceBandTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCostsWithProgressTracker() {
-    for (int n = 0; n < 10; n++) {
-      ProgressTracker<Permutation> originalTracker = new ProgressTracker<Permutation>();
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n, 20);
-      AcceptanceBandSampling<Permutation> ch =
-          new AcceptanceBandSampling<Permutation>(h, originalTracker);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      assertTrue(originalTracker == tracker);
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsSplit() {
     for (int n = 0; n < 10; n++) {
       IntProblem problem = new IntProblem();
@@ -243,28 +169,6 @@ public class AcceptanceBandTests extends SharedTestStochasticSampler {
       assertEquals(1, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsSplit() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n, 20);
-      AcceptanceBandSampling<Permutation> chOriginal = new AcceptanceBandSampling<Permutation>(h);
-      AcceptanceBandSampling<Permutation> ch = chOriginal.split();
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       Permutation p = solution.getSolution();
       assertEquals(n, p.length());
       tracker = new ProgressTracker<Permutation>();
@@ -453,31 +357,6 @@ public class AcceptanceBandTests extends SharedTestStochasticSampler {
   }
 
   @Test
-  public void testWithDoubleCostsBetaMax() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n, 20);
-      AcceptanceBandSampling<Permutation> ch = new AcceptanceBandSampling<Permutation>(h, 1.0);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize();
-      assertEquals(2, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
   public void testWithIntCostsBetaMin() {
     for (int n = 0; n < 10; n++) {
       IntProblem problem = new IntProblem();
@@ -496,31 +375,6 @@ public class AcceptanceBandTests extends SharedTestStochasticSampler {
       assertEquals(2, ch.getTotalRunLength());
       assertEquals((n + 1) * n / 2, solution.getCost());
       assertEquals((n + 1) * n / 2, tracker.getCost());
-      tracker = new ProgressTracker<Permutation>();
-      ch.setProgressTracker(tracker);
-      assertTrue(tracker == ch.getProgressTracker());
-    }
-  }
-
-  @Test
-  public void testWithDoubleCostsBetaMin() {
-    for (int n = 0; n < 10; n++) {
-      DoubleProblem problem = new DoubleProblem();
-      DoubleHeuristic h = new DoubleHeuristic(problem, n, 20);
-      AcceptanceBandSampling<Permutation> ch = new AcceptanceBandSampling<Permutation>(h, 0.0);
-      assertEquals(0, ch.getTotalRunLength());
-      assertTrue(problem == ch.getProblem());
-      ProgressTracker<Permutation> tracker = ch.getProgressTracker();
-      SolutionCostPair<Permutation> solution = ch.optimize();
-      assertEquals(1, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
-      Permutation p = solution.getSolution();
-      assertEquals(n, p.length());
-      solution = ch.optimize();
-      assertEquals(2, ch.getTotalRunLength());
-      assertEquals((n + 1) * n / 2, solution.getCostDouble(), 1E-10);
-      assertEquals((n + 1) * n / 2, tracker.getCostDouble(), 1E-10);
       tracker = new ProgressTracker<Permutation>();
       ch.setProgressTracker(tracker);
       assertTrue(tracker == ch.getProgressTracker());
